Se anula requerimiento de editar ALIAS, finalizado editar datos del usuario
