+++ digitalmoneyhouse/src/app/perfil/page.tsx
@@ -1,49 +1,11 @@
 "use client";
 import Menu from "@/components/common/menu/Menu";
 import HeaderDashboard from "@/components/common/navbar/HeaderDashboard";
-import AccountInfo from "@/components/perfil/AccountInfo";
 import PaymentGestion from "@/components/perfil/PaymentGestion";
-import accountAPI from "@/services/account/account.service";
-import userApi from "@/services/users/user.service";
-import { AccountData } from "@/types/account.types";
-import { UserType } from "@/types/users.types";
-import { useSession } from "next-auth/react";
-import { useEffect, useState } from "react";
-import ContainerData from "@/components/perfil/ContainerData";
+import ContainerDataUser from "@/components/perfil/user/ContainerDataUser";
+import ContainterDataAccount from "@/components/perfil/account/ContainterDataAccount";
 
 const page = () => {
-	const { data: session, status, update } = useSession();
-	const [userData, setUserData] = useState<UserType>({
-		dni: 0,
-		email: "",
-		firstname: "",
-		lastname: "",
-		password: "",
-		phone: "",
-	});
-	const [accountData, setAccountData] = useState<AccountData>({
-		alias: "",
-		available_amount: 0,
-		cvu: "",
-		id: 0,
-		user_id: 0,
-	});
-	const getDataUser = async () => {
-		if (session?.user?.token) {
-			const getAccount = await accountAPI.getMyAccount(`${session.user.token}`);
-			const res = await userApi.getUserData(
-				`${session.user.token}`,
-				getAccount.user_id
-			);
-			console.log(res);
-			setUserData(res);
-			setAccountData(getAccount);
-		}
-	};
-	useEffect(() => {
-		getDataUser();
-	}, [session, userData]);
-
 	return (
 		<>
 			<HeaderDashboard/>
@@ -53,9 +15,9 @@ const page = () => {
 						<Menu />
 					</div>
 					<div className="flex flex-col space-y-4 items-center py-5 md:w-8/12 lg:w-9/12 xl:w-10/12">
-						<ContainerData user={userData}/>
+						<ContainerDataUser/>
 						<PaymentGestion />
-						<AccountInfo accountData={accountData} />
+						<ContainterDataAccount/>
 					</div>
 				</div>
 			</main>
